backend/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
flask-cors==4.0.2
python-dotenv==1.0.1
requests==2.26.0
